Factor out renderActivityGroups.
